requirements.txt
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-material==1.4.3
django-phonenumber-field==2.2.0
idna==2.8
phonenumbers==8.10.5
PyJWT==1.7.1
PySocks==1.6.8
pytz==2018.9
requests==2.21.0
six==1.12.0
twilio==6.24.0
urllib3==1.24.1
django-cripsy-forms==1.7.2
django-bootstrap4==4.0.0.8
